[refactor] useHistoryFeature 불필요한 useEffect에 대한 리팩토링
- 함수 혹은 변수가 불리는 곳에 따라 useHistoryFeature, useHistoryMap으로 분리
